WebCrawler
This program will assess the status of the links (anchor tags) on a website (designed to be used on adobe.com but can be used on many other sites as well). This program also utilizes the Prawn gem to create a PDF file while displays the website that was crawled, the number of links found on the website, and lists the URLs in a clickable format and includes their link type and status (organized by status: valid/invalid/redirected). Also demonstrates practical usage of regular expressions to identify the anchors, as well as their response codes, link type and URL. Internal URLs are also concatinated into a searchable format.